Add the possibility to compensate an external disturbance torques
Add a copy of each constructors of `MotionConstrCommon`, `MotionConstr`
and `MotionSpringConstr` that includes an additional reference to an
`Eigen::VectorXd` that contains an estimation or the exact value of the
external disturbance torques to be compensated for.
`MotionConstrCommon` set a `bool` depending on which constructor is
called to know if external torques should be accounted for.
- This is required to know if the reference is properly initialized.
This value is later used in the `computeMatrix` and `computeTorque`
function of `MotionConstrCommon` if external torques needs to be
compensated.